Commit #2

+++ Jan_21_LW_1_T_1/Jan_21_LW_1_T_1/Program.cs
@@ -2,12 +2,12 @@
 
 namespace Jan_21_LW_1_T_1
 {
-    public delegate double DelegateA(double x, double y, double z);
+    public delegate int DelegateA(int x, int y, int z);
     class Program
     {
         static void Main(string[] args)
         {
-            DelegateA delA = delegate (double x, double y, double z) { return(x + y + z) / 3; };
+            DelegateA delA = delegate (int x, int y, int z) { return(x + y + z) / 3; };
             Console.WriteLine(delA(5, 5, 27));
             Console.ReadKey();
         }
